refactor(entities): align client fetchers with shared api
