+++ src/entities/brand/api/client/getBrands.js
@@ -1,4 +1,4 @@
-import { fetchJson } from '@/shared/api/fetchJson';
+import { fetchJson } from '@/shared/api';
 
 export async function getBrands({ signal } = {}) {
   return fetchJson('/api/brands', { signal });
