+++ src/entities/product/api/client/getProductById.js
@@ -1,4 +1,4 @@
-import { fetchJson } from '@/shared/api/fetchJson';
+import { fetchJson } from '@/shared/api';
 
 export async function getProductById({ id, signal } = {}) {
   if (id == null) throw new Error('getProductById: id is required');
